+++ lib/modules/security/security_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: security
-// Updated at: 2026-05-31 18:14:13
-// Commit: test(security): add unit tests for security business logic
+// Updated at: 2026-06-01 11:31:52
+// Commit: fix(security): prevent double tap duplicate action in security
 
 class securityComponentState {
-  final String commitRef = '1656';
-  final String lastMessage = 'test(security): add unit tests for security business logic';
-  final DateTime timestamp = DateTime.parse('2026-05-31T18:14:13+05:00');
+  final String commitRef = '1668';
+  final String lastMessage = 'fix(security): prevent double tap duplicate action in security';
+  final DateTime timestamp = DateTime.parse('2026-06-01T11:31:52+05:00');
 }